Hash when you change the password & Edit form to appear only to the Owner

+++ src/controllers/videoController.js
@@ -24,9 +24,16 @@ export const watch = async (req, res) => {
 // 유저가 getEdit로 올 때, 우린 편집용 form을 render해줄 거고,
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  // video.owner는 object형태이고, _id는 string형태라서, String()로 바꿔준다.
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 }; // edit.pug를 렌더링한다.
@@ -35,10 +42,17 @@ export const getEdit = async (req, res) => {
 // postEdit은 route로부터 id를 얻어와서 /videos/id 페이지로 redirect해준다.
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
-  const video = await Video.exists({ _id: id });
+  const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  // video.owner는 object형태이고, _id는 string형태라서, String()로 바꿔준다.
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -91,7 +105,21 @@ export const postUpload = async (req, res) => {
 // <Delete Video>
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  // video.owner는 object형태이고, _id는 string형태라서, String()로 바꿔준다.
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
+  user.videos.splice(user.videos.indexOf(id), 1);
+  user.save();
   return res.redirect("/");
 };
